String transform method moved to DBHelper
