+++ src/main/java/cz/honzakasik/offensesindex/database/DBManager.java
@@ -3,6 +3,7 @@ package cz.honzakasik.offensesindex.database;
 import com.healthmarketscience.sqlbuilder.*;
 import com.healthmarketscience.sqlbuilder.dbspec.basic.*;
 import cz.honzakasik.offensesindex.database.populator.DBPopulator;
+import javafx.collections.ObservableList;
 
 import java.sql.*;
 import java.util.logging.Level;
@@ -152,9 +153,9 @@ public class DBManager {
         return null;
     }
 
-    public ResultSet getAllCities() {
+    public ObservableList<String> getAllCities() {
         String citiesQuery = new SelectQuery(true)
                 .addCustomColumns(driversTable.findColumn(CITY)).validate().toString();
-        return executeSQL(citiesQuery);
+        return DBHelper.transformStringData(executeSQL(citiesQuery), CITY);
     }
 }
