+++ src/main/java/cz/honzakasik/offensesindex/policemen/PolicemenTabController.java
@@ -1,7 +1,7 @@
 package cz.honzakasik.offensesindex.policemen;
 
+import cz.honzakasik.offensesindex.DialogHelper;
 import cz.honzakasik.offensesindex.Helper;
-import cz.honzakasik.offensesindex.database.DBHelper;
 import cz.honzakasik.offensesindex.database.DBManager;
 import cz.honzakasik.offensesindex.database.DepartmentsDBManager;
 import cz.honzakasik.offensesindex.database.PolicemenDBManager;
@@ -12,8 +12,6 @@ import javafx.stage.Stage;
 import org.controlsfx.validation.ValidationResult;
 import org.controlsfx.validation.ValidationSupport;
 import org.controlsfx.validation.Validator;
-
-import java.sql.ResultSet;
 
 /**
  * Created by Jan Kasik on 14.11.15.
@@ -32,7 +30,7 @@ public class PolicemenTabController {
         policemenDBManager = new PolicemenDBManager(dbManager);
         DepartmentsDBManager departmentsDBManager = new DepartmentsDBManager(dbManager);
         policemenTable.setItems(policemenDBManager.getAllPolicemen());
-        policemenDepartmentSelector.setItems(Helper.transformDepartmentSelectorData(departmentsDBManager.getAllDepartmentsNames()));
+        policemenDepartmentSelector.setItems(departmentsDBManager.getAllDepartmentsNames());
         bindPolicemenValidationSupport();
     }
 
@@ -52,7 +50,7 @@ public class PolicemenTabController {
         ObservableList<PolicemanTableItem> result = policemenDBManager
                 .getPolicemenFromDepartmentWithinYear(policemenDepartmentSelector.getValue(), year);
         if (result.isEmpty()) {
-            Helper.displayNothingFoundError((Stage)findPolicemenButton.getScene().getWindow());
+            DialogHelper.displayNothingFoundError((Stage)findPolicemenButton.getScene().getWindow());
             policemenYear.clear();
             policemenDepartmentSelector.setValue(null);
             policemenTable.setItems(policemenDBManager.getAllPolicemen());
